增加是否删除war解压文件的选项

+++ easy-deploy.sh
@@ -2,10 +2,29 @@
 
 # author : mingjian_hou
 # desc : shut down the tomcat, clear the cache and startup it, please put this script in /tomcat/bin/
+# args : 
+#        -s : save(not delete) the webapps/xxx(unzip war) when run this script.
 
 # DEFINE
 tomcat_name="tomcat";
 war_name="xxxxx"
+
+# THE ARGS
+isDeleteUnzipWar=true
+
+# INIT ARGS, DEAL THE -s ARGS
+while getopts d opt
+do 
+    case "$opt" in
+    s)
+        isDeleteUnzipWar=false
+    ;;
+    \?)
+        echo "Invalid option: -$OPTARG"
+        exit
+    ;;
+    esac
+done
 
 # get tomcat pid and judge should shutdown or not
 tomcat_id=$(ps -aux | grep -w .*"$tomcat_name".*Bootstrap.* | grep -v 'grep' | awk '{print $2}')
@@ -16,7 +35,7 @@ else
     # run shut down 
     ./shutdown.sh
     echo "===== closing the tomcat, waiting ... ====="
-    sleep 3
+    sleep 6
     
     # recheck the tomcat pid exist or not
     tomcat_id=$(ps -aux | grep -w .*"$tomcat_name".*Bootstrap.* | grep -v 'grep' | awk '{print $2}')
@@ -27,8 +46,13 @@ else
 fi
 
 # remove the tomcat cache
-rm -rf ../webapps/$war_name
+if [ $isDeleteUnzipWar = true ]; then
+    rm -rf ../webapps/$war_name
+    echo "===== delete the unzip war file ====="
+    sleep 1
+fi
 rm -rf ../work/Catalina/*
+rm -rf ../conf/Catalina/*
 rm -rf ../temp/*
 # restart tomcat
 ./startup.sh
@@ -36,4 +60,3 @@ rm -rf ../temp/*
 # tail the log
 cur_date=$(date +%F)
 tail -f ../logs/catalina.out.${cur_date}
-
